+++ R/jjbarstats.h.R
@@ -8,6 +8,7 @@ jjbarstatsOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
         initialize = function(
             dep = NULL,
             group = NULL,
+            grvar = NULL,
             excl = TRUE,
             originaltheme = FALSE, ...) {
 
@@ -33,6 +34,14 @@ jjbarstatsOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
                     "nominal"),
                 permitted=list(
                     "factor"))
+            private$..grvar <- jmvcore::OptionVariable$new(
+                "grvar",
+                grvar,
+                suggested=list(
+                    "ordinal",
+                    "nominal"),
+                permitted=list(
+                    "factor"))
             private$..excl <- jmvcore::OptionBool$new(
                 "excl",
                 excl,
@@ -44,17 +53,20 @@ jjbarstatsOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
 
             self$.addOption(private$..dep)
             self$.addOption(private$..group)
+            self$.addOption(private$..grvar)
             self$.addOption(private$..excl)
             self$.addOption(private$..originaltheme)
         }),
     active = list(
         dep = function() private$..dep$value,
         group = function() private$..group$value,
+        grvar = function() private$..grvar$value,
         excl = function() private$..excl$value,
         originaltheme = function() private$..originaltheme$value),
     private = list(
         ..dep = NA,
         ..group = NA,
+        ..grvar = NA,
         ..excl = NA,
         ..originaltheme = NA)
 )
@@ -63,6 +75,7 @@ jjbarstatsResults <- if (requireNamespace('jmvcore')) R6::R6Class(
     inherit = jmvcore::Group,
     active = list(
         todo = function() private$.items[["todo"]],
+        plot2 = function() private$.items[["plot2"]],
         plot = function() private$.items[["plot"]]),
     private = list(),
     public=list(
@@ -86,10 +99,23 @@ jjbarstatsResults <- if (requireNamespace('jmvcore')) R6::R6Class(
                     "originaltheme")))
             self$add(jmvcore::Image$new(
                 options=options,
-                name="plot",
-                title="Bar Chart",
+                name="plot2",
+                title="Bar Chart Splitted",
                 width=800,
                 height=600,
+                renderFun=".plot2",
+                requiresData=TRUE,
+                clearWith=list(
+                    "dep",
+                    "group",
+                    "grvar",
+                    "direction",
+                    "originaltheme"),
+                visible="(grvar)"))
+            self$add(jmvcore::Image$new(
+                options=options,
+                name="plot",
+                title="Bar Chart",
                 renderFun=".plot",
                 requiresData=TRUE,
                 clearWith=list(
@@ -132,11 +158,13 @@ jjbarstatsBase <- if (requireNamespace('jmvcore')) R6::R6Class(
 #' @param data The data as a data frame.
 #' @param dep .
 #' @param group .
+#' @param grvar .
 #' @param excl .
 #' @param originaltheme .
 #' @return A results object containing:
 #' \tabular{llllll}{
 #'   \code{results$todo} \tab \tab \tab \tab \tab a html \cr
+#'   \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
 #'   \code{results$plot} \tab \tab \tab \tab \tab an image \cr
 #' }
 #'
@@ -145,6 +173,7 @@ jjbarstats <- function(
     data,
     dep,
     group,
+    grvar,
     excl = TRUE,
     originaltheme = FALSE) {
 
@@ -153,18 +182,22 @@ jjbarstats <- function(
 
     if ( ! missing(dep)) dep <- jmvcore::resolveQuo(jmvcore::enquo(dep))
     if ( ! missing(group)) group <- jmvcore::resolveQuo(jmvcore::enquo(group))
+    if ( ! missing(grvar)) grvar <- jmvcore::resolveQuo(jmvcore::enquo(grvar))
     if (missing(data))
         data <- jmvcore::marshalData(
             parent.frame(),
             `if`( ! missing(dep), dep, NULL),
-            `if`( ! missing(group), group, NULL))
+            `if`( ! missing(group), group, NULL),
+            `if`( ! missing(grvar), grvar, NULL))
 
     for (v in dep) if (v %in% names(data)) data[[v]] <- as.factor(data[[v]])
     for (v in group) if (v %in% names(data)) data[[v]] <- as.factor(data[[v]])
+    for (v in grvar) if (v %in% names(data)) data[[v]] <- as.factor(data[[v]])
 
     options <- jjbarstatsOptions$new(
         dep = dep,
         group = group,
+        grvar = grvar,
         excl = excl,
         originaltheme = originaltheme)